+++ app/Jobs/Server/FirewallRules/RemoveServerNetworkRule.php
@@ -45,7 +45,9 @@ class RemoveServerNetworkRule implements ShouldQueue
             $serverService->getService(SystemService::FIREWALL, $this->serverNetworkRule->server)->removeServerNetworkRule($this->serverNetworkRule->server->ip);
         });
 
-        $this->serverNetworkRule->delete();
+        if ($this->wasSuccessful()) {
+            $this->serverNetworkRule->delete();
+        }
 
         throw new ServerCommandFailed($this->getCommandErrors());
     }
